Optionally upload archives to Azure Blob Storage post-download
This updates the script to upload the migration archives to Azure
Blob Storage as they are downloaded from GitHub, deleting them
from local storage as we go.

This allows the script to be safely run with large numbers of
repos on a machine that doesn't have gigantic amounts of storage,
since storage is effectively offloaded to Azure.